Refactor to put the getWeather async function in its own module
