+++ src/index.js
@@ -1,14 +1,5 @@
 import { convertTempToCelsius, convertTempToFahrenheit } from './tempMethods';
-
-const getWeather = async (city) => {
-  const response = await fetch(`https://api.openweathermap.org/data/2.5/weather?q=${city}&appid=3a907874d0e5af1a38cfac90720e9f08`);
-
-  if (response.status !== 200) {
-    throw new Error('Please enter full city name. Cannot fetch the data');
-  }
-  const data = await response.json();
-  return data;
-};
+import getWeather from './fetchWeather';
 
 const displayWeatherInfo = document.getElementById('displayWeatherCard');
